# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Evolution GO - whatsmeow",
        "title": "Evolution GO",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/instance/all": {
            "get": {
                "description": "Get all instances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Get all instances",
                "responses": {
                    "200": {
                        "description": "All instances",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/connect": {
            "post": {
                "description": "Connect to instance with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Connect to instance",
                "parameters": [
                    {
                        "description": "Instance data",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.ConnectStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance connected successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/create": {
            "post": {
                "description": "Creates a new instance with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Create a new instance",
                "parameters": [
                    {
                        "description": "Instance data",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.CreateStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance created successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/delete/{instanceId}": {
            "delete": {
                "description": "Delete instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Delete instance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instance Id",
                        "name": "instanceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/disconnect": {
            "post": {
                "description": "Disconnect from instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Disconnect from instance",
                "responses": {
                    "200": {
                        "description": "Instance disconnected successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/logout": {
            "delete": {
                "description": "Logout from instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Logout from instance",
                "responses": {
                    "200": {
                        "description": "Instance logged out successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/pause": {
            "post": {
                "description": "Pause instance connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Pause instance",
                "responses": {
                    "200": {
                        "description": "Instance paused successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/resume": {
            "post": {
                "description": "Resume a previously paused instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Resume instance",
                "responses": {
                    "200": {
                        "description": "Instance resumed successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/instance/pair": {
            "post": {
                "description": "Request pairing code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Request pairing code",
                "parameters": [
                    {
                        "description": "Instance data",
                        "name": "instance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.PairStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pairing code",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Error on validation",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/instance/qr": {
            "get": {
                "description": "Get instance QR code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instance"
                ],
                "summary": "Get instance QR code",
                "responses": {
                    "200": {
                        "description": "Instance QR code",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "github_com_Zapbox-API_evolution-go_pkg_instance_service.ConnectStruct": {
            "type": "object",
            "properties": {
                "immediate": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "subscribe": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "webhookUrl": {
                    "type": "string"
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_instance_service.CreateStruct": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "proxy": {
                    "$ref": "#/definitions/github_com_Zapbox-API_evolution-go_pkg_instance_service.ProxyConfig"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_instance_service.PairStruct": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string"
                },
                "subscribe": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_Zapbox-API_evolution-go_pkg_instance_service.ProxyConfig": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}